GCA-167 create repository layer

+++ gym-common/src/main/java/com/gymcommon/workload/MonthlyWorkloadRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 
 public record MonthlyWorkloadRequest(
         @NotBlank String username,
-        @NotBlank Integer month,
-        @NotBlank Integer year
+        @NotBlank Short month,
+        @NotBlank Short year
 ) {
 }
