+++ gym-common/src/main/java/com/gymcommon/workload/MonthlyWorkloadResponse.java
@@ -2,9 +2,9 @@ package com.gymcommon.workload;
 
 public record MonthlyWorkloadResponse(
         String username,
-        int year,
-        int month,
-        int totalHours,
-        int totalMinutes
+        Short year,
+        Short month,
+        Long totalHours,
+        Long totalMinutes
 ) {
 }
